backend/requirements.txt
# Framework web
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# LLM y embeddings
ollama
sentence-transformers

# Vector store
faiss-cpu

# Procesamiento de documentos
pdfplumber
python-docx
PyPDF2

# Utilidades
python-multipart
python-dotenv
